validator_cloud_function/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.1
cloudevents==1.11.1
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
Flask==3.1.1
functions-framework==3.8.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.82.1
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
PyPDF2==3.0.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
watchdog==6.0.0
Werkzeug==3.1.3
google-cloud-storage